changed type nonce

+++ gameplay/queue/chore.go
@@ -315,6 +315,7 @@ func (chore *Chore) FinishWithWinResult(ctx context.Context, winResult WinResult
 
 	winResult.GameResult.Question = "you allow us to take your address?"
 	winResult.GameResult.Transaction.Value = cryptoutils.WeiToEthereum(winResult.Value).String()
+	// TODO: send before you allow address
 	winResult.GameResult.Transaction.UDTContract.Address = chore.config.UDTContract.Address
 	if err := winResult.Client.WriteJSON(http.StatusOK, winResult.GameResult); err != nil {
 		chore.log.Error("could not write json", ChoreError.Wrap(err))
@@ -349,7 +350,8 @@ func (chore *Chore) FinishWithWinResult(ctx context.Context, winResult WinResult
 			}
 		}
 
-		if winResult.GameResult.Transaction, err = chore.currencywaitlist.Create(ctx, user.ID, *winResult.Value, request.Nonce); err != nil {
+		// TODO: int64 > int if 32bit os
+		if winResult.GameResult.Transaction, err = chore.currencywaitlist.Create(ctx, user.ID, *winResult.Value, int64(request.Nonce)); err != nil {
 			chore.log.Error("could not create item of currencywaitlist", ChoreError.Wrap(err))
 			return
 		}
